requirements.txt
sqlalchemy==1.3.13
jupyter==1.0.0
matplotlib==3.1.3
numpy==1.18.1
pandas==1.0.1
psycopg2==2.8.4
qwikidata==0.4.0
scikit-learn==0.22.1
yoyo-migrations==7.0.0
wikitables==0.5.4